+++ src/core_read.cpp
@@ -7,6 +7,7 @@
 #include "core.h"
 #include "script/script.h"
 #include "serialize.h"
+#include "streams.h"
 #include "univalue/univalue.h"
 #include "util.h"
 #include "utilstrencodings.h"
