+++ src/test/alert_tests.cpp
@@ -10,6 +10,7 @@
 #include "data/alertTests.raw.h"
 
 #include "serialize.h"
+#include "streams.h"
 #include "util.h"
 #include "utilstrencodings.h"
 #include "version.h"
