+++ src/txmempool.h
@@ -12,6 +12,8 @@
 #include "core.h"
 #include "sync.h"
 
+class CAutoFile;
+
 inline bool AllowFree(double dPriority)
 {
     // Large (in bytes) low-priority (new, small-coin) transactions
